requirements.txt
asgiref==3.3.1
coverage==5.4
Django==3.1.6
django-debug-toolbar==3.2
django-test-plus==1.4.0
isort==5.7.0
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
